fix(export): réduit la duplication de code détectée par SonarCloud
Extrait le parsing de la fenêtre de dates dans parseExportDateWindow()
(fetchRepresentations.ts) et réutilise errorSchema pour les réponses
400/500 de la doc OpenAPI, au lieu de dupliquer inline la structure déjà
présente ailleurs dans le fichier.
