+++ packages/app/src/modules/export/fetchRepresentations.ts
@@ -5,6 +5,40 @@ import { and, eq, gte, lt } from "drizzle-orm";
 import { toNumber } from "~/modules/public-api";
 import { db } from "~/server/db";
 import { companies, representationDeclarations } from "~/server/db/schema";
+import { exportDeclarationsQuerySchema } from "./schemas";
+
+function getNextDate(date: string): string {
+	const d = new Date(`${date}T00:00:00Z`);
+	d.setUTCDate(d.getUTCDate() + 1);
+	return d.toISOString().slice(0, 10);
+}
+
+export function parseExportDateWindow(
+	request: Request,
+): { date_begin: string; dateEnd: string } | Response {
+	const url = new URL(request.url);
+	const parsed = exportDeclarationsQuerySchema.safeParse({
+		date_begin: url.searchParams.get("date_begin") ?? undefined,
+		date_end: url.searchParams.get("date_end") ?? undefined,
+	});
+
+	if (!parsed.success) {
+		return Response.json(
+			{
+				error:
+					"Paramètres invalides. 'date_begin' est requis, format YYYY-MM-DD.",
+				details: parsed.error.issues,
+			},
+			{ status: 400 },
+		);
+	}
+
+	const { date_begin } = parsed.data;
+	return {
+		date_begin,
+		dateEnd: parsed.data.date_end ?? getNextDate(date_begin),
+	};
+}
 
 async function fetchSubmittedRepresentationsRows(
 	dateBegin: string,
